Larger Refactor
I was playing around with CSS in JS solutions Tailwind,  Sass I enjoy leveraging bootstrap class based styling. This led me into what other ways I might be able to manage the data modal that is readable and had the commenting and syntax. So this is a take on using yaml as the one source of truth data modal for Oday so far testing has been good performance from trimming the outside JSON library for internal Deno readTextFile as well as using encoding yaml to translate the files which is also a base Deno method.